Some populate files clean up
