--- apps/smart-forms-app/src/functions/populate-functions/PrePopQueryPopulateFunctions.ts
@@ -1,40 +1,0 @@
-/*
- * Copyright 2023 Commonwealth Scientific and Industrial Research
- * Organisation (CSIRO) ABN 41 687 119 230.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { Questionnaire } from 'fhir/r4';
-import { isLaunchContext, isSourceQuery } from './populateTypePredicates.ts';
-import type { LaunchContext } from './launchContext.interface.ts';
-import type { SourceQuery } from './sourceQueries.interface.ts';
-
-export function getLaunchContexts(questionnaire: Questionnaire): LaunchContext[] {
-  if (questionnaire.extension && questionnaire.extension.length > 0) {
-    return questionnaire.extension.filter((extension) =>
-      isLaunchContext(extension)
-    ) as LaunchContext[];
-  }
-
-  return [];
-}
-
-// get source query references
-export function getSourceQueries(questionnaire: Questionnaire): SourceQuery[] {
-  if (questionnaire.extension && questionnaire.extension.length > 0) {
-    return questionnaire.extension.filter((extension) => isSourceQuery(extension)) as SourceQuery[];
-  }
-
-  return [];
-}
