+++ apps/smart-forms-app/src/functions/populateFunctions/typePredicates.ts
@@ -16,11 +16,11 @@
  */
 
 import type { Extension } from 'fhir/r4';
-import type { LaunchContext } from './launchContext.interface.ts';
+import type { LaunchContext } from '../../interfaces/populateInterfaces/launchContext.interface.ts';
 import type {
   QuestionnaireLevelXFhirQueryVariable,
   SourceQuery
-} from './sourceQueries.interface.ts';
+} from '../../interfaces/populateInterfaces/sourceQueries.interface.ts';
 
 export function isLaunchContext(extension: Extension): extension is LaunchContext {
   const hasLaunchContextName =
